requirements.txt
discord.py
emoji
